Move pipelines to separated project
