+++ divrec/utils/__init__.py
@@ -1,5 +1,5 @@
-from .divrec_logger import get_logger
+from .divrec_logger import get_logger, get_file_handler
 from .score_with_reduction import ScoreWithReduction
 
 
-__all__ = ["ScoreWithReduction", "get_logger"]
+__all__ = ["ScoreWithReduction", "get_logger", "get_file_handler"]
