+++ divrec_experiments/experiments/matrix_factorization_ml100k/main.py
@@ -1,4 +1,4 @@
-from divrec_experiments.pipeline import Pipeline
+from divrec_pipelines.pipeline import Pipeline
 from divrec_experiments.experiments.matrix_factorization_ml100k.src import load_dataset, test_model, train_model
