+++ divrec_experiments/experiments/matrix_factorization_ml100k/stages.yaml
@@ -1,20 +1,23 @@
-load_dataset:
-  path: /Users/alexey.tsyplov/Projects/diversity-recommendations/divrec_experiments/data/ml-100k
-  train_size: 0.7
-  test_size: 0.2
-train_model:
-  train_max_sampled: 100
-  train_batch_size: 100
-  validation_max_sampled: 100
-  validation_batch_size: 100
-  embedding_dim: 100
-  lr: 0.001
-  epochs: 10
-  model_path: /Users/alexey.tsyplov/Projects/diversity-recommendations/divrec_experiments/experiments/matrix_factorization_ml100k/workdir/model.pth
+meta:
   logfile: /Users/alexey.tsyplov/Projects/diversity-recommendations/divrec_experiments/experiments/matrix_factorization_ml100k/workdir/pipeline.log
-test_model:
-  model_path: /Users/alexey.tsyplov/Projects/diversity-recommendations/divrec_experiments/experiments/matrix_factorization_ml100k/workdir/model.pth
-  embedding_dim: 100
-  max_sampled: -1
-  test_scores_filepath: /Users/alexey.tsyplov/Projects/diversity-recommendations/divrec_experiments/experiments/matrix_factorization_ml100k/workdir/scores.json
-  logfile: /Users/alexey.tsyplov/Projects/diversity-recommendations/divrec_experiments/experiments/matrix_factorization_ml100k/workdir/pipeline.log
+stages:
+  load_dataset:
+    path: /Users/alexey.tsyplov/Projects/diversity-recommendations/divrec_experiments/data/ml-100k
+    train_size: 0.7
+    test_size: 0.2
+  train_model:
+    train_max_sampled: 100
+    train_batch_size: 100
+    validation_max_sampled: 100
+    validation_batch_size: 100
+    embedding_dim: 100
+    lr: 0.001
+    epochs: 10
+    model_path: /Users/alexey.tsyplov/Projects/diversity-recommendations/divrec_experiments/experiments/matrix_factorization_ml100k/workdir/model.pth
+    logfile: /Users/alexey.tsyplov/Projects/diversity-recommendations/divrec_experiments/experiments/matrix_factorization_ml100k/workdir/pipeline.log
+  test_model:
+    model_path: /Users/alexey.tsyplov/Projects/diversity-recommendations/divrec_experiments/experiments/matrix_factorization_ml100k/workdir/model.pth
+    embedding_dim: 100
+    max_sampled: -1
+    test_scores_filepath: /Users/alexey.tsyplov/Projects/diversity-recommendations/divrec_experiments/experiments/matrix_factorization_ml100k/workdir/scores.json
+    logfile: /Users/alexey.tsyplov/Projects/diversity-recommendations/divrec_experiments/experiments/matrix_factorization_ml100k/workdir/pipeline.log
