--- divrec_experiments/pipeline/pipeline.py
@@ -1,29 +1,0 @@
-import os
-from typing import List
-
-import click
-import yaml
-
-from divrec_experiments.pipeline import Stage
-
-
-class Pipeline:
-    def __init__(self, stages: List[Stage]):
-        self.stages = stages
-
-    def run(self):
-        @click.command()
-        @click.argument("config_path")
-        def wrapper(config_path) -> None:
-            with open(os.path.abspath(config_path), mode="r") as file:
-                config = yaml.safe_load(file)
-
-            head, tail = self.stages[0], self.stages[1:]
-
-            head.config = config.get(head.name, dict())
-            output = head()
-            for stage in tail:
-                stage.config = config.get(stage.name, dict())
-                output = stage(output)
-
-        return wrapper()
